apps/api/src/requirements.txt
camelot-py[base]===0.10.1
